+++ Proyecto_LDPII/src/main/java/com/cibertec/app/repository/ContratoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.cibertec.app.entity.Contrato;
 
 @Repository
-public interface ContratoRepository extends JpaRepository<Contrato, Integer> {
+public interface ContratoRepository extends JpaRepository<Contrato, Long> {
 
 }
