+++ Proyecto_LDPII/src/main/java/com/cibertec/app/service/EventoService.java
@@ -7,12 +7,12 @@ import java.util.List;
 public interface EventoService {
     public List<Evento> getAllEvento();
 
-    public List<Evento> getEventoByCliente(Integer idCliente);
+    public List<Evento> getEventoByCliente(Long idCliente);
 
     public Evento crearEvento(Evento evento);
 
-    public String deleteEvento(Integer idEvento);
+    public String deleteEvento(Long idEvento);
 
-    public Evento buscarEventoById(Integer id);
+    public Evento buscarEventoById(Long id);
 
 }
